feat: Run a command while serving a directory via HTTP.
Usage: run-with-server [OPTIONS] CMD...

Run a command while serving a directory via HTTP.

This command first starts a simple HTTP server on a free port, runs the
given command and finally shuts down the HTTP server.

The server will serve the directory given with --serve-dir (current
directory by default) and all its subdirectories. It will find a free TCP
port that is at least start_port to run the server. If --verbose, the server
will output an access log and some more details.

The remaining arguments specify the command to run and its arguments. The
special string {port} will be replaced with the port number used by the
server. If --shell is given, the system shell will be used to run the
command, in this case only one non-option argument should be given.

The exit code will be passed on from the command.